requirements.txt
aiohttp>=2.2.5
alphabet_detector
beautifulsoup4
bleach
cachetools
configparser
currencyconverter
emoji
feedparser
future
git+https://github.com/python-telegram-bot/ptbcontrib.git
gitpython
gpytranslate
markdown2
multicolorcaptcha
pillow>=4.2.0
psutil
psycopg2-binary
pyYAML>=5.1.2
pyrate-limiter
python-telegram-bot==13.11
requests
spamwatch
speedtest-cli
sqlalchemy==1.4.11
telethon
ujson
gtts
wikipedia
SibylSystem
pytimeparse
lxml
pycryptodome
base58~=2.1.1
redis
bs4
pytz
gtts
wikipedia
